reorg to new signalComponentHelpers util
